use uint16_t instead of int

+++ Inc/led.h
@@ -5,7 +5,7 @@
 #ifndef PLAYALONG_LED_H
 #define PLAYALONG_LED_H
 
-void on_strum(int interval);
+void on_strum(uint16_t interval);
 void on_chord(char* chord);
 void on_wait(char* chord);
